Update UserSettingModel and refine download-mod event
The `UserSettingModel` has been extended with `isActive` and `isValid` fields. The download-mod event and `personal-setting.component` have been updated to allow selecting an active user setting. Also, a `DownloadModel` interface has been added, which includes the download URL and install path to assist the process of downloading and installing mods.