Add functionality to remove user settings
This update implements a new listener in the user-setting event class. This listener responds to a 'user-settings-remove' event, removing specific user settings entries from the store. It also updates the personal setting component to utilize this new functionality. The component will now trigger the 'user-settings-remove' event when a user chooses to remove their stored settings, and it will automatically refresh the user settings being displayed.