Add ModSearch component and update AppComponent
A ModSearch component has been created for handling the mod search functionality which was previously part of AppComponent. The new component ensures the separation of concerns and brings more clarity to the codebase. Existing styling and HTML structure were moved from AppComponent to this new component. Hence, the AppComponent has been updated accordingly to use this new ModSearch component.