Update mod-search component with improved layout and mod management
This commit revises the mod-search component to better manage mod items. The updated layout now allows for more content with the removal and addition of mod items streamlined. Moreover, a new functionality to check if a mod item is already in the modList has been integrated which subsequently governs the visibility of add and delete buttons. CSS classes have been updated resulting in better content visibility and uniformity across the application.