Add support for 7z archive extraction
This update introduces support for extracting files from 7z archives alongside the existing support for ZIP files. This extends the functionality of the file unzipping event, and requires an additional dependency on the 'node-7z' library. Modifying the mod list service also removes hard-coded mod list and uses an empty array instead. It also replaces adm-zip with 7z for mod extraction in order to support more types of mods.