Update app version and GitHub action
The application version was updated in package.json from "0.0.1" to "0.0.3". Also, in .github/workflows/main.yml, the version for the action-electron-builder was specified to "v1.6.0" and an argument to always publish was added. This indicates that the app should be released after building whenever a commit is tagged with a version.