Remove TODO.md file
The TODO.md file which contained the task list for the project has been removed. This could mean that all tasks have been completed, or they might be migrated to a different tracking system, but more likely, the tracking of tasks has been moved directly into issues or cards on the project management board.