Update UI, modify package metadata and add window event handling
This commit updates the user interface for improved readability and user experience. The description of the package has been modified along with the title tag of the application. Moreover, window event handling has been introduced for the main-application-start file to manage 'window-close', 'window-minimize', and 'window-maximize' events. Several style changes and class additions are made in app.component.scss for better structure and design.