Update user-settings management and enhance mod installation check
The code updates include a revamp on user-settings management inclusive of modifications to 'open-directory' and 'user-settings-update' events. The 'is-already-installed' directive is also tweaked to enhance the mod installation checking process. It also includes minor additions to the user-settings service such as the removal and addition of user-settings based on directory.