Add VirusTotal scan to release workflow
A new workflow has been added that triggers a VirusTotal scan on each new release. The scan, which runs on the latest version of Windows, uses the VirusTotal API key stored in the repository secrets to scan all executable (.exe) files in the release package. Results are then appended to the release body.