Refactor interface and implement a sorting feature
The changes in this commit adjust the user interface and add a new feature to sort generic mod lists. Icon and toolbar adjustments were made in the generic mod list and in the side navigation of the app component. The core function to load data in the generic mod list component can now sort the mods by 'downloads'. These modifications improve usability and enable easier browsing of mods.