Add external path opening feature in settings
The commit introduces a new feature in the personal settings component that allows opening of an external path. A new button is added to the general settings UI, which upon clicking opens the external path on the system. The feature is implemented using the Electron shell to carry out the operation.