Add support for downloading from Dropbox links
The code has been updated to handle Dropbox download links properly. A new function `isDropBox` has been added to check if the link is from Dropbox. If the link ends with `&dl=0`, indicating it's a shared link that displays the file in the Dropbox interface, it's replaced to end with `&dl=1` which directly downloads the file.