Update mod list and improve archive handling functions
Updated the mod list in the ModListService, replacing an outdated mod with a new one. Enhanced the file-unzip.event to handle RAR archives in addition to other archive types. These improvements include verifying if an archive contains a single DLL file and following a 'happy path' approach to extraction. Added 'node-unrar-js' as a new dependency, supporting the extraction of RAR files.