Improve mod download functionality
The mod download process has been improved by adding a check if the download form exists on the page before the browser is closed. Additionally, an error handling process has been improved to ensure the browser is closed on receiving a download error. In a minor cleanup process, unnecessary comments have been removed.