Add Qodana configuration and refactor webpack setup
A Qodana configuration file (qodana.yaml) was added for the application analysis. This includes settings for executing shell commands before Qodana execution and enabling specific inspections. In addition, the existing webpack configuration in angular.webpack.js file was refactored for styling and indentation improvements. The changes do not alter the functionality of the file.