Add external link button to mod cards
The code updates feature the addition of a clickable icon on mod cards that will open an external link. This new feature is made visible on mouse hover and its styling is also defined. The changes aim to improve the user interaction by providing a more direct, intuitive way to access the external content related to the mods.