Update mod download and installation behavior
Modified the condition for displaying the delete button in mod-card component. Injected Download Service in generic-mod-list and mod-search components to reflect if a download and install action is in progress. This allows to update the tooltips for the add button accordingly. Increased the download timeout limit in the download-mod.event.