Add informative messages for empty mod lists
This commit adds display messages for cases where the server or client mods lists in the Instance Overview component are empty. This offers more clarity to the user on the status of their installed mods. Additionally, the commit adds new mod names to the is-already-installed directive and updates a function call in the app component.