Add extension functionality to mod cards
The update introduces the ability to extend mod cards in the Mod List component menu. This change helps to improve the visibility of information for each mod. The mod extension capability is activated through a new button on each card, and it includes a new animation to smoothly transition between the "normal" and "extended" states. A new property, "extended", has also been added to the Mod model to track whether a card is extended or not.