Fixing (temp) overlapping button, fixing some name mappings
The flex property in the generic-mod-list component styling has been modified for better alignment. Also, the names of the "SVM" mods in both the alternativeServerModNames and alternativeClientModNames objects have been updated to 'Server Value Modifier [SVM]' from 'KMC Server Value Modifier (SVM)'. This change improves clarity and consistency in the mod names presented to the user.