Update application behavior and implement notification system
This commit mainly includes significant updates to the application's behavior according to platform and implements a notification system for mod updates. Unnecessary comments were removed from 'main-window.ts'. Additional tray functionality and setting of the app user model ID were enabled for windows platform in 'main-application-start.ts'. In 'download.service.ts', the fileId was added to the FileUnzipEvent. Additional mods were added for initial configuration in 'mod-list.service.ts'. Furthermore, a new notification system was implemented which notifies users when a new version of a mod is published.