Add supported Aki version display in mod cards
Added a new feature to show the supported Aki version for each mod in the mod cards. The different colours represent the status of the Aki version for each mod, added a label with a specific colour code for each mod according to its supported Aki version. This information is picked up from the mod's file URL. Additionally, moved the .image-aki-version SCSS styling from the generic-mod-list to the global styles.