Update URLs based on environment in aki-search.service
The key change in this commit is an update to the modSearchUrl and modUrl in the aki-search.service.ts file, such that they now depend on whether the environment is production or not. This allows the application to switch between the production server and local server easily, enhancing the development process.