Add warning for unsupported mod installations
Modified components and services to include a warning for mod installations that are not supported. This update includes a new boolean property 'notSupported' in the mod model to flag unsupported mod installations. Furthermore, updates were made across various components to incorporate the warning that is shown to the user whilst dealing with unsupported mod installations.