Refactor loadData call and add empty list handling
The loadData method call in generic-mod-list.component.ts was moved out of the sortField setter and into the ngOnInit lifecycle hook. Additionally, the `tags` input property was refactored to potentially accept undefined value. And for handling the case when accumulatedModList is empty and tags input is true, a user message has been added in generic-mod-list.component.html file.