Add method to fetch current version in Configuration Service
The Configuration Service now calls and processes the current version separately from the current tags. This was done so as to have a cleaner structure and an easier debugging process in case of any arising issues. The relevant changes have also been reflected in the main and mod_list files.