Make 'supportedAkiVersion' property required and sort mods by version
The 'supportedAkiVersion' property in the 'mod' data model has been made required for more consistent data handling. Also, a modification has been made to the search component logic to sort the mods by the 'supportedAkiVersion'. This allows mods to be displayed in a more organized and predictable manner to users.