Add MediaFire link check and error handling in file operations
This commit adds an additional check to determine if a download link is a MediaFire link, in order to prevent potentially problematic downloads. It also adds error handling in various parts of file operations, including more robust checks for the existence of specific files. In addition, superfluous actions in the download events have been commented out to improve performance.