Add tracking for completion of download process
A new behavior subject has been added in 'download.service.ts' to track when the download process is completed. In 'app.component.ts', subscribed to this behavior subject to get updates about the status of download. Disabled a button in 'instance-overview.component.html' by commenting out the functionality for future revisions.