Add loading spinner to mod search component
Code modifications were implemented to add a loading spinner on the mod search component for asynchronous operations. Changes allow users to identify when the app is fetching data during a search operation, thereby improving user experience. The loading spinner gets displayed when there are no mod items available and the system is making an API call to fetch data.