Update dependencies to latest stable versions
This commit updates a variety of dependencies used in the application, including "@jsverse/transloco", "material-symbols", "@ngneat/spectator", "electron" and more. This is part of ongoing maintenance of the project to keep the system using up-to-date libraries, taking advantage of new features or fixes, and keeping compatibility with newer systems or environments where the application will be run.