Add handling for PowerShell issues
The updates made in the code are aimed at introducing PowerShell error handling in the application. The main updates involved adding a new property `isPowerShellIssue` to differentiate it from other errors. This new field is used in various services, models, and event handlers. A visual indicator for the error was also added in the `personal-setting.component`. In the case of a PowerShell error, it prevents loading client-mods instead of marking the directory as invalid.