Refactor mod toggling and get disabled server mods
This commit includes a major refactor of the mod toggling functionality. It extracts the logic from the component to a service and adds checks for active instances. It also optimizes the server mods event to check and return disabled server mods along with their details for use in the client. A minor change at modifying the 'modOriginalName' fields from 'mod' to 'mod.name' has been performed as well.