Replace Aki references with Spt in the app
This commit updates various parts of the application, replacing references to 'aki' with 'spt'. This includes code changes in the TypeScript and HTML files, renaming of model interfaces, and updating paths used for file unzipping and downloading. Please note that these changes are part of a larger refactoring to accommodate the naming changes in the underlying systems and models.